Move request_body_file pos arg to normal option
Currently 'request_body_file' argument is implemented as a positional
argument. A positional argument is usually used for required parameter,
but 'request_body_file' is optional. It is better to implement an
optional argument as normal argparse option.
The parameter also takes multiple files, but it is rarely useful.

This commit moves 'request_body_file' positional argument to
an optional argument named '--input-file'.